BCDA-3578 moved api files and updated tests
