+++ bcda/api/v1/api_test.go
@@ -1,4 +1,4 @@
-package web
+package v1
 
 import (
 	"context"
@@ -22,6 +22,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
+	api "github.com/CMSgov/bcda-app/bcda/api"
 	"github.com/CMSgov/bcda-app/bcda/auth"
 	"github.com/CMSgov/bcda-app/bcda/constants"
 	"github.com/CMSgov/bcda-app/bcda/database"
@@ -49,16 +50,24 @@ type RequestParams struct {
 }
 
 var origDate string
+var origBBCert string
+var origBBKey string
 
 func (s *APITestSuite) SetupSuite() {
 	s.reset = testUtils.SetUnitTestKeysForAuth() // needed until token endpoint moves to auth
 	origDate = os.Getenv("CCLF_REF_DATE")
 	os.Setenv("CCLF_REF_DATE", time.Now().Format("060102 15:01:01"))
 	os.Setenv("BB_REQUEST_RETRY_INTERVAL_MS", "10")
+	origBBCert = os.Getenv("BB_CLIENT_CERT_FILE")
+	os.Setenv("BB_CLIENT_CERT_FILE", "../../../shared_files/decrypted/bfd-dev-test-cert.pem")
+	origBBKey = os.Getenv("BB_CLIENT_KEY_FILE")
+	os.Setenv("BB_CLIENT_KEY_FILE", "../../../shared_files/decrypted/bfd-dev-test-key.pem")
 }
 
 func (s *APITestSuite) TearDownSuite() {
 	os.Setenv("CCLF_REF_DATE", origDate)
+	os.Setenv("BB_CLIENT_CERT_FILE", origBBCert)
+	os.Setenv("BB_CLIENT_KEY_FILE", origBBKey)
 	s.reset()
 }
 
@@ -346,7 +355,8 @@ func bulkEOBRequestHelper(endpoint, since string, s *APITestSuite) {
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -382,7 +392,8 @@ func bulkEOBRequestInvalidSinceFormatHelper(endpoint, since string, s *APITestSu
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -428,7 +439,8 @@ func bulkEOBRequestNoBeneficiariesInACOHelper(endpoint string, s *APITestSuite) 
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -457,7 +469,8 @@ func bulkEOBRequestMissingTokenHelper(endpoint string, s *APITestSuite) {
 }
 
 func bulkEOBRequestNoQueueHelper(endpoint string, s *APITestSuite) {
-	qc = nil
+	// qc := nil
+	api.SetQC(nil)
 
 	acoID := acoUnderTest
 	jobCount, err := s.getJobCount(acoID)
@@ -519,7 +532,8 @@ func bulkPatientRequestHelper(endpoint, since string, s *APITestSuite) {
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -553,7 +567,8 @@ func bulkPatientRequestInvalidSinceFormatHelper(endpoint, since string, s *APITe
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -598,7 +613,8 @@ func bulkCoverageRequestHelper(endpoint string, requestParams RequestParams, s *
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -632,7 +648,8 @@ func bulkCoverageRequestInvalidSinceFormatHelper(endpoint, since string, s *APIT
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -674,7 +691,8 @@ func bulkCoverageRequestInvalidSinceDateHelper(endpoint, since string, s *APITes
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -716,7 +734,8 @@ func bulkCoverageRequestInvalidOutputHelper(endpoint, outputFormat string, s *AP
 	}
 	defer pgxpool.Close()
 
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	handler := http.HandlerFunc(handlerFunc)
 	handler.ServeHTTP(s.rr, req)
@@ -938,7 +957,7 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 	req := httptest.NewRequest("GET", requestUrl.String(), nil)
 	rctx := chi.NewRouteContext()
 	req = req.WithContext(context.WithValue(req.Context(), chi.RouteCtxKey, rctx))
-	resourceTypes, err := validateRequest(req)
+	resourceTypes, err := api.ValidateRequest(req)
 	assert.Nil(s.T(), resourceTypes)
 	assert.Equal(s.T(), responseutils.Error, err.Issue[0].Severity)
 	assert.Equal(s.T(), responseutils.Exception, err.Issue[0].Code)
@@ -947,7 +966,7 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 
 	requestParams := RequestParams{}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 3, len(resourceTypes))
 	for _, t := range resourceTypes {
@@ -958,7 +977,7 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 
 	requestParams = RequestParams{resourceType: "ExplanationOfBenefit,Patient"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 2, len(resourceTypes))
 	for _, t := range resourceTypes {
@@ -969,7 +988,7 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 
 	requestParams = RequestParams{resourceType: "Coverage,Patient"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 2, len(resourceTypes))
 	for _, t := range resourceTypes {
@@ -980,28 +999,28 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 
 	requestParams = RequestParams{resourceType: "ExplanationOfBenefit"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 1, len(resourceTypes))
 	assert.Contains(s.T(), resourceTypes, "ExplanationOfBenefit")
 
 	requestParams = RequestParams{resourceType: "Patient"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 1, len(resourceTypes))
 	assert.Contains(s.T(), resourceTypes, "Patient")
 
 	requestParams = RequestParams{resourceType: "Coverage"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), err)
 	assert.Equal(s.T(), 1, len(resourceTypes))
 	assert.Contains(s.T(), resourceTypes, "Coverage")
 
 	requestParams = RequestParams{resourceType: "Practitioner"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), resourceTypes)
 	assert.Equal(s.T(), responseutils.Error, err.Issue[0].Severity)
 	assert.Equal(s.T(), responseutils.Exception, err.Issue[0].Code)
@@ -1010,7 +1029,7 @@ func validateRequestHelper(endpoint string, s *APITestSuite) {
 
 	requestParams = RequestParams{resourceType: "Patient,Patient"}
 	_, _, req = bulkRequestHelper(endpoint, requestParams)
-	resourceTypes, err = validateRequest(req)
+	resourceTypes, err = api.ValidateRequest(req)
 	assert.Nil(s.T(), resourceTypes)
 	assert.Equal(s.T(), responseutils.Error, err.Issue[0].Severity)
 	assert.Equal(s.T(), responseutils.Exception, err.Issue[0].Code)
@@ -1024,9 +1043,9 @@ func bulkRequestHelper(endpoint string, testRequestParams RequestParams) (string
 	var group string
 
 	if endpoint == "Patient" {
-		handlerFunc = bulkPatientRequest
+		handlerFunc = BulkPatientRequest
 	} else if endpoint == "Group/all" {
-		handlerFunc = bulkGroupRequest
+		handlerFunc = BulkGroupRequest
 		group = groupAll
 	}
 
@@ -1053,7 +1072,7 @@ func bulkRequestHelper(endpoint string, testRequestParams RequestParams) (string
 func (s *APITestSuite) TestJobStatusInvalidJobID() {
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%s", "test"), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", "test")
@@ -1078,7 +1097,7 @@ func (s *APITestSuite) TestJobStatusJobDoesNotExist() {
 	jobID := "1234"
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%s", jobID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", jobID)
@@ -1111,7 +1130,7 @@ func (s *APITestSuite) TestJobStatusPending() {
 	req, err := http.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 	assert.Nil(s.T(), err)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1138,7 +1157,7 @@ func (s *APITestSuite) TestJobStatusInProgress() {
 
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1166,7 +1185,7 @@ func (s *APITestSuite) TestJobStatusFailed() {
 
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1206,7 +1225,7 @@ func (s *APITestSuite) TestJobStatusCompleted() {
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 	req.TLS = &tls.ConnectionState{}
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1265,7 +1284,7 @@ func (s *APITestSuite) TestJobStatusCompletedErrorFileExists() {
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 	req.TLS = &tls.ConnectionState{}
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1324,7 +1343,7 @@ func (s *APITestSuite) TestJobStatusExpired() {
 
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1353,7 +1372,7 @@ func (s *APITestSuite) TestJobStatusNotExpired() {
 
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1379,7 +1398,7 @@ func (s *APITestSuite) TestJobStatusArchived() {
 
 	req := httptest.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 
-	handler := http.HandlerFunc(jobStatus)
+	handler := http.HandlerFunc(JobStatus)
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1402,7 +1421,7 @@ func (s *APITestSuite) TestServeData() {
 	rctx.URLParams.Add("fileName", "test.ndjson")
 	req = req.WithContext(context.WithValue(req.Context(), chi.RouteCtxKey, rctx))
 
-	handler := http.HandlerFunc(serveData)
+	handler := http.HandlerFunc(ServeData)
 	handler.ServeHTTP(s.rr, req)
 
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
@@ -1414,7 +1433,7 @@ func (s *APITestSuite) TestMetadata() {
 	req := httptest.NewRequest("GET", "/api/v1/metadata", nil)
 	req.TLS = &tls.ConnectionState{}
 
-	handler := http.HandlerFunc(metadata)
+	handler := http.HandlerFunc(Metadata)
 	handler.ServeHTTP(s.rr, req)
 
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
@@ -1423,7 +1442,7 @@ func (s *APITestSuite) TestMetadata() {
 func (s *APITestSuite) TestGetVersion() {
 	req := httptest.NewRequest("GET", "/_version", nil)
 
-	handler := http.HandlerFunc(getVersion)
+	handler := http.HandlerFunc(GetVersion)
 	handler.ServeHTTP(s.rr, req)
 
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
@@ -1448,7 +1467,7 @@ func (s *APITestSuite) TestJobStatusWithWrongACO() {
 	req, err := http.NewRequest("GET", fmt.Sprintf("/api/v1/jobs/%d", j.ID), nil)
 	assert.Nil(s.T(), err)
 
-	handler := auth.RequireTokenJobMatch(http.HandlerFunc(jobStatus))
+	handler := auth.RequireTokenJobMatch(http.HandlerFunc(JobStatus))
 
 	rctx := chi.NewRouteContext()
 	rctx.URLParams.Add("jobID", fmt.Sprint(j.ID))
@@ -1466,7 +1485,7 @@ func (s *APITestSuite) TestJobStatusWithWrongACO() {
 func (s *APITestSuite) TestHealthCheck() {
 	req, err := http.NewRequest("GET", "/_health", nil)
 	assert.Nil(s.T(), err)
-	handler := http.HandlerFunc(healthCheck)
+	handler := http.HandlerFunc(HealthCheck)
 	handler.ServeHTTP(s.rr, req)
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
 }
@@ -1483,7 +1502,7 @@ func (s *APITestSuite) TestHealthCheckWithBadDatabaseURL() {
 	os.Setenv("DATABASE_URL", "not-a-database")
 	req, err := http.NewRequest("GET", "/_health", nil)
 	assert.Nil(s.T(), err)
-	handler := http.HandlerFunc(healthCheck)
+	handler := http.HandlerFunc(HealthCheck)
 	handler.ServeHTTP(s.rr, req)
 	assert.Equal(s.T(), http.StatusBadGateway, s.rr.Code)
 }
@@ -1496,7 +1515,7 @@ func (s *APITestSuite) TestAuthInfoDefault() {
 	// set provider to bogus value and make sure default (alpha) is retrieved
 	auth.SetProvider("bogus")
 	req := httptest.NewRequest("GET", "/_auth", nil)
-	handler := http.HandlerFunc(getAuthInfo)
+	handler := http.HandlerFunc(GetAuthInfo)
 	handler.ServeHTTP(s.rr, req)
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
 	respMap := make(map[string]string)
@@ -1518,7 +1537,7 @@ func (s *APITestSuite) TestAuthInfoAlpha() {
 	// set provider to alpha and make sure alpha is retrieved
 	auth.SetProvider("alpha")
 	req := httptest.NewRequest("GET", "/_auth", nil)
-	handler := http.HandlerFunc(getAuthInfo)
+	handler := http.HandlerFunc(GetAuthInfo)
 	handler.ServeHTTP(s.rr, req)
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
 	respMap := make(map[string]string)
@@ -1540,7 +1559,7 @@ func (s *APITestSuite) TestAuthInfoOkta() {
 	// set provider to okta and make sure okta is retrieved
 	auth.SetProvider("okta")
 	req := httptest.NewRequest("GET", "/_auth", nil)
-	handler := http.HandlerFunc(getAuthInfo)
+	handler := http.HandlerFunc(GetAuthInfo)
 	handler.ServeHTTP(s.rr, req)
 	assert.Equal(s.T(), http.StatusOK, s.rr.Code)
 	respMap := make(map[string]string)
@@ -1588,7 +1607,8 @@ func makeConnPool(s *APITestSuite) *pgx.ConnPool {
 	if err != nil {
 		s.T().Error(err)
 	}
-	qc = que.NewClient(pgxpool)
+	qc := que.NewClient(pgxpool)
+	api.SetQC(qc)
 
 	return pgxpool
 }
